pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "xrpl-camp"
version = "0.1.0"
description = "Learn the XRPL diary — wallet, payment, verification, certificate in one sitting."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "mcp-tool-shop", email = "64996768+mcp-tool-shop@users.noreply.github.com" },
]
keywords = ["xrpl", "web3", "education", "ledger", "testnet"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Education",
]
dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "xrpl-py>=2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
]

[project.scripts]
xrpl-camp = "xrpl_camp.cli:app"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["xrpl_camp"]
